feat: refactor tx details popup
